+++ client/src/components/forms/index.js
@@ -1,4 +1,3 @@
-export { default as OptionsForm } from "./OptionsForm";
 export { default as CopyRuleForm } from "./CopyRuleForm";
 export { default as EdgeNetworkingForm } from "./EdgeNetworkingForm";
 export { default as F5Form } from "./F5Form";