+++ unit-tests/state/options.test.js
@@ -156,5 +156,243 @@ describe("options", () => {
         "it should have correct craig version"
       );
     });
+    describe("options.schema", () => {
+      let craig;
+      beforeEach(() => {
+        craig = newState();
+      });
+      it("should return region groups when fs cloud", () => {
+        assert.deepEqual(
+          craig.options.region.groups({ fs_cloud: true }),
+          ["eu-de", "eu-gb", "us-east", "us-south"],
+          "it should return correct groups"
+        );
+      });
+      it("should return region groups when not fs cloud", () => {
+        assert.deepEqual(
+          craig.options.region.groups({ fs_cloud: false }),
+          [
+            "au-syd",
+            "br-sao",
+            "ca-tor",
+            "eu-de",
+            "eu-gb",
+            "jp-osa",
+            "jp-tok",
+            "us-east",
+            "us-south",
+          ],
+          "it should return correct groups"
+        );
+      });
+      it("should have invalid prefix when none prefix", () => {
+        assert.isTrue(craig.options.prefix.invalid({}), "it should be invalid");
+      });
+      it("should have invalid prefix when longer than 16 characters", () => {
+        assert.isTrue(
+          craig.options.prefix.invalid({ prefix: "looooooooooooooooooong" }),
+          "it should be invalid"
+        );
+      });
+      it("should have invalid prefix when bad value", () => {
+        assert.isTrue(
+          craig.options.prefix.invalid({ prefix: "@@@" }),
+          "it should be invalid"
+        );
+      });
+      it("should set power vs regions to [] and return region on input change", () => {
+        let data = {
+          region: "us-south",
+        };
+        assert.deepEqual(
+          craig.options.region.onInputChange(data),
+          "us-south",
+          "it should return correct region"
+        );
+        assert.deepEqual(
+          data,
+          {
+            region: "us-south",
+            power_vs_zones: [],
+          },
+          "it should set power vs zones"
+        );
+      });
+      it("should render public and private endpoints in title case", () => {
+        assert.deepEqual(
+          craig.options.endpoints.onRender({ endpoints: "public-and-private" }),
+          "Public and Private",
+          "it should return endpoints"
+        );
+      });
+      it("should disable dynamic subnets toggle when advanced subnets is true", () => {
+        assert.isTrue(
+          craig.options.dynamic_subnets.disabled(
+            {},
+            {
+              craig: {
+                store: { json: { _options: { advanced_subnets: true } } },
+              },
+            }
+          ),
+          "it should be disabled"
+        );
+      });
+      it("should disable dynamic subnets toggle when advanced subnets not found", () => {
+        assert.isFalse(
+          craig.options.dynamic_subnets.disabled(
+            {},
+            {
+              craig: {
+                store: { json: { _options: {} } },
+              },
+            }
+          ),
+          "it should be disabled"
+        );
+      });
+      it("should change enable_power_vs from false to true", () => {
+        let data = { enable_power_vs: false };
+        craig.options.enable_power_vs.onStateChange(data);
+        assert.deepEqual(
+          data,
+          {
+            enable_power_vs: true,
+          },
+          "it should change value"
+        );
+      });
+      it("should change enable_power_vs from true to false and reset zones", () => {
+        let data = { enable_power_vs: true };
+        craig.options.enable_power_vs.onStateChange(data);
+        assert.deepEqual(
+          data,
+          {
+            enable_power_vs: false,
+            power_vs_zones: [],
+          },
+          "it should change value"
+        );
+      });
+      it("should change power_vs_high_availability from false to true", () => {
+        let data = { power_vs_high_availability: false };
+        craig.options.power_vs_high_availability.onStateChange(data);
+        assert.deepEqual(
+          data,
+          {
+            power_vs_high_availability: true,
+            power_vs_zones: ["dal12", "wdc06"],
+          },
+          "it should change value"
+        );
+      });
+      it("should change power_vs_high_availability from true to false and reset zones", () => {
+        let data = { power_vs_high_availability: true };
+        craig.options.power_vs_high_availability.onStateChange(data);
+        assert.deepEqual(
+          data,
+          {
+            power_vs_high_availability: false,
+            power_vs_zones: [],
+          },
+          "it should change value"
+        );
+      });
+      it("should hide power_vs_high_availability when not using power vs", () => {
+        assert.isTrue(
+          craig.options.power_vs_high_availability.hideWhen({
+            enable_power_vs: false,
+          }),
+          "it should be hidden"
+        );
+      });
+      it("should return correct invalid text when region does not have power vs zones", () => {
+        assert.deepEqual(
+          craig.options.power_vs_zones.invalidText({ region: "invalid" }),
+          "The region invalid does not have any available Power VS zones",
+          "it should return correct invalid text"
+        );
+      });
+      it("should be invalid when region does not have power vs zones", () => {
+        assert.isTrue(
+          craig.options.power_vs_zones.invalid({
+            region: "invalid",
+            enable_power_vs: true,
+            power_vs_zones: [],
+          }),
+          "it should be invalid"
+        );
+      });
+      it("should be invalid when region does not have power vs zones", () => {
+        assert.isTrue(
+          craig.options.power_vs_zones.invalid({
+            region: "invalid",
+            enable_power_vs: true,
+            power_vs_zones: ["dal12", "wdc06"],
+          }),
+          "it should be invalid"
+        );
+      });
+      it("should be valid when region does not have power vs zones and power vs not enabled", () => {
+        assert.isFalse(
+          craig.options.power_vs_zones.invalid({
+            region: "invalid",
+            enable_power_vs: false,
+            power_vs_zones: [],
+          }),
+          "it should be invalid"
+        );
+      });
+      it("should return correct invalid text when does have power vs zones", () => {
+        assert.deepEqual(
+          craig.options.power_vs_zones.invalidText({ region: "us-south" }),
+          "Select at least one Availability Zone",
+          "it should return correct invalid text"
+        );
+      });
+      it("should force update zones on power vs high availability toggle", () => {
+        assert.isTrue(
+          craig.options.power_vs_zones.forceUpdateKey({
+            power_vs_high_availability: true,
+          }),
+          "it should return value"
+        );
+      });
+      it("should return correct power_vs_zones groups when power_vs_high_availability true", () => {
+        assert.deepEqual(
+          craig.options.power_vs_zones.groups({
+            power_vs_high_availability: true,
+          }),
+          ["dal12", "wdc06"],
+          "it should return correct groups"
+        );
+      });
+      it("should return correct power_vs_zones groups when power_vs_high_availability false", () => {
+        assert.deepEqual(
+          craig.options.power_vs_zones.groups({
+            power_vs_high_availability: false,
+            region: "ca-tor",
+          }),
+          ["tor01"],
+          "it should return correct groups"
+        );
+      });
+      it("should return tags when no value", () => {
+        assert.deepEqual(
+          craig.options.tags.onInputChange({}),
+          [],
+          "it should return empty array"
+        );
+      });
+      it("should return tags when value", () => {
+        assert.deepEqual(
+          craig.options.tags.onInputChange({
+            tags: "hello,world",
+          }),
+          ["hello", "world"],
+          "it should return empty array"
+        );
+      });
+    });
   });
 });