+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - SCC v1 Terraform has been deprecated. SCC v1 resources can still be managed from `/form/securityComplianceCenter`
 - Reafactored Routing Tables and Routing Table Routes Form
+- OptionsForm has been removed and now is generated using the DynamicForm component
 
 ### Features